Add files via upload
This project is a safety-focused smart helmet using Arduino UNO. It detects alcohol using an MQ-3 sensor, monitors accidents via a vibration sensor, and checks if the helmet is worn using a tactile/reed switch. When triggered, it sends an SMS with GPS location to a predefined mobile number via a GSM module. You can even change the emergency number by sending an SMS command. It's a low-cost, beginner-friendly project ideal for engineering students.